snip pkg/util/strings dependency
