+++ pkg/genericapiserver/api/BUILD
@@ -35,7 +35,6 @@ go_library(
         "//pkg/runtime:go_default_library",
         "//pkg/runtime/schema:go_default_library",
         "//pkg/util/errors:go_default_library",
-        "//pkg/util/strings:go_default_library",
         "//vendor:github.com/emicklei/go-restful",
     ],
 )
