+++ pkg/util/strings/strings_test.go
@@ -95,49 +95,3 @@ func TestIsVowel(t *testing.T) {
 		}
 	}
 }
-
-func TestGetArticleForNoun(t *testing.T) {
-	type args struct {
-	}
-	tests := []struct {
-		noun    string
-		padding string
-		want    string
-	}{
-		{
-			noun:    "Frog",
-			padding: " ",
-			want:    " a ",
-		},
-		{
-			noun:    "frogs",
-			padding: " ",
-			want:    " ",
-		},
-		{
-			noun:    "apple",
-			padding: "",
-			want:    "an",
-		},
-		{
-			noun:    "Apples",
-			padding: " ",
-			want:    " ",
-		},
-		{
-			noun:    "Ingress",
-			padding: " ",
-			want:    " an ",
-		},
-		{
-			noun:    "Class",
-			padding: " ",
-			want:    " a ",
-		},
-	}
-	for _, tt := range tests {
-		if got := GetArticleForNoun(tt.noun, tt.padding); got != tt.want {
-			t.Errorf("%q. GetArticleForNoun() = %v, want %v", tt.noun, got, tt.want)
-		}
-	}
-}
